put user identifier inside jwt token
The reason, to put userId inside JWT token, is to get user identifier without request repository. This technique is usefull because only one request is made to repository (only when the token is created).
Moreover this is profitable when microservices are used because isn’t necessary request the gateway to know the userId for the current user. Neither is necessary duplicate all about user on every microservice