Updated the JDL section
Before, there was only one page with all the information regarding the JDL was
written.

For someone who doesn't have any knowledge about the JDL, this is not easy to
get basic information.
For someone with more knowledge, and looking to know more, one has to navigate
through things already seen.

From the contributor's viewpoint, updating the single page isn't fun:
  - the contributor would have to pinpoint the place where to make changes
  - and make sure this place is the right one

From the reviewer's viewpoint, the same issues arise.

This is how this commit changes things:
  - a "getting started" page, for easy information about basic knowledge to get started
  - for each JDL notion (entity, field, enum, etc.): a page has been made
    - with examples
    - with precisions about the syntax
  - also for troubleshooting