Update security.md to reflect network and platform
The KeyCloak server running as a daemon in docker-compose has some specific configuration "gotchas" that are not obvious on first pass. Additionally, KeyCloak may not properly start or fail outright on Apple Silicon. Include instructions for building the image locally to address the issues that appear to be related directly to Rosetta 2 (Compatibility Mode).