docs: rewrite API first guide with detailed implementation and samples (#1335)
* docs: rewrite API first guide with detailed implementation and samples

Re-ordered concepts to introduce concepts provided by JHipster with details on the out of the box
configuration of the openapi-generator and the use of the Delegate Pattern. Explicitly callout the
use of the OAI's Pet Store example with links to samples.

Added notes on JHipster's technical structure and the use of ArchUnit to prepare the reader for the
implementation of the API operations.

Rewrite and editdocs for a detailed introduction of API first development

NOTE: removed references to the Maven IDEA and Eclipse plugins. Both have been retired and should be
removed from JHipster.

Added detailed examples of the implementation of NativeWebRequest and responses

Lots of examples and details added to both the basic and detailed implementation with mock request
responses prior to implementation with the NativeWebRequest.

Add sample api.yml used in guide and build out initial content

The content is largely complete and ready for review. Add links and the sample api.yml document used
throughout the guide.

* docs: add details on obtaining an authorization token for API tests

JHipster defaults the `/api/**` behind authentication via the SecurityConfigurationFilter chain,
e.g., `.authenticated()`. It is possible to craft a JWT Bearer token to test the API using the JWT
Secret generated by JHipster during creation.

* docs: adjust the permalink property of the yaml frontmatter

Test the desired URL to help Jekyll with redirects

* docs: clean up grammar, punctuation, and phrasing for additional clarity

.

* docs: use Service over Component

Update docs to reflect OpenAPI implementation recommendation.

* docs: remove note on openapi-client subgenerator

The openapi-client subgenerator is no longer maintained after v8.

```shell
 % jhipster openapi-client
WARNING! Since JHipster v8, the jhipster command will not use the locally installed generator-jhipster.
    If you want to execute the locally installed generator-jhipster, run: npx jhipster
FATAL!

This sub-generator is no longer maintained and has been removed from JHipster starting from v8.
If you are interested in maintaining this sub-generator fill a feature request on the JHipster bug tracker.
```

* docs: clarify the choice of API First and the generator

Generically state the configuration of the project for API First tooling and link to the OpenAPI
Generator page for more information.

* refactor: move sample api.yml to options directory

Linking static file for reference.

* refactor: move api sample to static directory

Following Docusaurus documentation. Create a directory in /static/<reference-file>/path

Update path.